Create MIPS Array Modifier
This repository contains a MIPS assembly program that takes a numerical input value from the user and performs operations on an array of two integers, A[2], based on the user's input. The program initializes the elements of the array to -2 and then implements logic similar to the provided C code to modify the array elements.

The logic implemented is as follows:

If the user input (i) is 0, the program sets A[0] to 0.
If the user input is 1 or 2, the program sets A[1] to 1.
If the user input is any other value, A[0] is set to -1.
After processing the user input, the program will print out the values of 'i' and the updated array 'A'.